Updated bot to sell food, and updated sellfood method to wait a bit longer between sales to allow tuna to be sold instead of banked.
The short delay before was causing other food to be sold too early, and then added to a solditems list which blocked them from being sold when the script was ready to sell that food type. This is now fixed by adding a longer delay and removing that solditems list.