Quick backup cos I almost killed her after hours of work.
Added Task class which will be the new root of everything that a script (BotMan) does.

Tasks will be able to be added, removed, and edited via the botMenu even during script operation.

I'm just not sure if we'll be able to edit the first one safely but I'll look into thread safety when I have time