Managed to get the task queue linked with the botmenu (FINALLY...)
The remove button also seems to work. Tomorrow will look at adding a settins panel for each script to load script-specific settings into.

Will also provide a reset button to restart the last saved task-set.

And a save task-set button which just saves the current task-queue as it is (unless it's empty). Then when load is pressed, this saved queue should be there (using ghostQueue as a backup)