Seems to have successfully loaded a live copy of the task list onto the botmenu, I can also remove tasks from it using the remove button on the menu.
The task library has some functionality now, maybe half complete? The library tab has a title, which

Made it so that everytime a Task is created, or added to the Task Man, the Task Library automatically inherits it, if it's not already contained in the library. Later I want to make this a dictionary for smoother referencing.