Seems to have successfully loaded a live copy of the task list onto the botmenu, I can also remove tasks from it using the remove button on the menu.
The task library has some functionality now, maybe half complete? The library tab has a title, which

Made it so that everytime a Task is created, or added to the Task Man, the Task Library automatically inherits it, if it's not already contained in the library. Later I want to make this a dictionary for smoother referencing.

Adjusted TaskMan's add() function so that it automatically updats the task library whenever someone submits a task (considering just making this happy when a task is created instead?)

There is also no need for an addUrgent method anymore, all Tasks now have an isUrgent field that can be toggled and is checked when being added to the taskman. If isUrgent, task is added to index 1 in the task queue to prevent thread locking errors hopefully (should be next in line?)