Added basic prototype for settings menu
Dev controls will be available to everyone until I setup a server and database.

added a bunch of getters/setters for the menu and had to rework a bunch of stuff to maintain privacy.

Made taskman a lot more private to make it more secure.

added lots of getters/setters to botman to compensate for lack of access to taskman now.

Moved list from botmenu to taskman to avoid confusion and keep like things together. Those lists are essentially the task lists and the task man is essentially a task list so it was very confusing having them split.