Added increment/decrements script index functions for navigation buttons later and easier ref.
Mainly just made changes to how the main loop works for the task menu to incorporate the two different loop types, script loops and task loops.

Adjusted getnextask() and getPreviousTask() ready for button attachment, now they cant go past the start or end of the queue and move the pointer for the list/display on call.