Fixed issues with bot menu not launching
Reworked task manager to have task loops and list loops.

Task loops: the number of execution loops for each task on the task-managers task-list.

List loops: the number of execution loops for the full set of tasks on the task-managers task-list

Added a task-library to the bot menu, this is a list of all created tasks and is populated automatically as each task in created in the bot menu.